FLAS-44 Move operation response to DTO
