+++ src/DTO/OperationResponse.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 
-namespace App\Service\Response;
+namespace App\DTO;
 
 
 /**
