+++ src/Service/Lesson/WordImageService.php
@@ -12,6 +12,8 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 
+//* ENTER POINT
+
 readonly class WordImageService implements WordImageServiceInterface
 {
     /**
